# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chat de Atendimento - API
  description: |
    API RESTful para sistema de atendimento ao cliente via WhatsApp.
    
    ## Funcionalidades Principais
    
    - 🔐 **Autenticação JWT** com SQL Server
    - 💬 **WhatsApp Business API** (Meta Cloud)
    - 🤖 **Chatbot IA** com Google Gemini
    - ⚡ **WebSocket** para tempo real
    - 📊 **Dashboard** com métricas
    
    ## Fluxo de Autenticação
    
    1. Faça login via `POST /auth/login`
    2. Use o token retornado no header `Authorization: Bearer <token>`
    3. Token expira em 24 horas
    
  version: 2.0.0
  contact:
    name: Suporte API
    email: api@empresa.com
    url: https://empresa.com/suporte
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Ambiente de Desenvolvimento
  - url: https://api.empresa.com/api/v1
    description: Ambiente de Produção
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Autenticar usuário
      description: |
        Autentica um usuário e retorna um token JWT.
        
        **Rate Limit:** 5 tentativas / 15 minutos por IP
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: usuario@email.com
              password: senha123
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - auth
      summary: Fazer logout
      description: Revoga o token JWT atual
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout bem-sucedido
        '401':
          description: Token inválido
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /auth/login
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: usuario@email.com
        password:
          type: string
          format: password
          minLength: 8
          example: senha123
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 86400
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: usuario@email.com
        nome:
          type: string
          example: João Silva
        role:
          type: string
          enum: [admin, atendente, supervisor, user]
          example: atendente
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        error_code:
          type: string
        field:
          type: string
security:
  - bearerAuth: []